httpd2: propagate defense errors across streams
Previously, if a defense mechanism triggered in response to one stream
in an HTTP/2 connection, that stream would shut down. In most cases this
is enough, since it looks like typical HTTP/2 attack software doesn't
handle stream reset very well, and this tends to stop further requests.

But smarter attack software could easily work around this.

So, this changes the connection-level dispatcher to notice a Defense
error on any single request, and propagate that to a hard shutdown of
the connection at the TCP level.